code clean, linting, search icon on keyboard, svg alignment
